merge feature_42128

+++ spec/models/enrollments/individual_market/open_enrollment_begin_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 require "#{Rails.root}/spec/shared_contexts/enrollment.rb"
 
@@ -26,8 +28,7 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
                             enrollment_kind: "open_enrollment",
                             kind: "individual",
                             submitted_at: start_on.prev_month,
-                            plan_id: plan.id,
-                      
+                            product_id: product.id
                             )
                           }
 
@@ -57,24 +58,23 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
       let(:params)                  { valid_params }
       let!(:benefit_coverage_period) { BenefitCoveragePeriod.new(**params) }
 
-      let(:active_individual_health_plan)       { FactoryBot.create(:active_individual_health_plan) }
-      let(:active_shop_health_plan)             { FactoryBot.create(:active_shop_health_plan) } 
-      let(:active_individual_dental_plan)       { FactoryBot.create(:active_individual_dental_plan) } 
-      let(:active_individual_catastophic_plan)  { FactoryBot.create(:active_individual_catastophic_plan) }
-      let(:active_csr_87_plan)                  { FactoryBot.create(:active_csr_87_plan) } 
-      let(:active_csr_00_plan)                  { FactoryBot.create(:active_csr_00_plan) }
+      let(:active_individual_health_product)       { FactoryBot.create(:active_individual_health_product) }
+      let(:active_shop_health_product)             { FactoryBot.create(:active_shop_health_product) }
+      let(:active_individual_dental_product)       { FactoryBot.create(:active_individual_dental_product) }
+      let(:active_individual_catastophic_product)  { FactoryBot.create(:active_individual_catastophic_product) }
+      let(:active_csr_87_product)                  { FactoryBot.create(:active_csr_87_product) }
+      let(:active_csr_00_product)                  { FactoryBot.create(:active_csr_00_product) }
 
-      let(:renewal_individual_health_plan)       { FactoryBot.create(:renewal_individual_health_plan) }
-      let(:renewal_shop_health_plan)             { FactoryBot.create(:renewal_shop_health_plan) } 
-      let(:renewal_individual_dental_plan)       { FactoryBot.create(:renewal_individual_dental_plan) } 
-      let(:renewal_individual_catastophic_plan)  { FactoryBot.create(:renewal_individual_catastophic_plan) }
-      let(:renewal_csr_87_plan)                  { FactoryBot.create(:renewal_csr_87_plan) } 
-      let(:renewal_csr_00_plan)                  { FactoryBot.create(:renewal_csr_00_plan) }
-      let!(:plan) { FactoryBot.create(:plan, :with_premium_tables, market: 'individual', metal_level: 'gold', active_year: TimeKeeper.date_of_record.year, hios_id: "11111111122302-01", csr_variant_id: "01")}
+      let(:renewal_individual_health_product)       { FactoryBot.create(:renewal_individual_health_product) }
+      let(:renewal_shop_health_product)             { FactoryBot.create(:renewal_shop_health_product) }
+      let(:renewal_individual_dental_product)       { FactoryBot.create(:renewal_individual_dental_product) }
+      let(:renewal_individual_catastophic_product)  { FactoryBot.create(:renewal_individual_catastophic_product) }
+      let(:renewal_csr_87_product)                  { FactoryBot.create(:renewal_csr_87_product) }
+      let(:renewal_csr_00_product)                  { FactoryBot.create(:renewal_csr_00_product) }
+      let!(:product) { FactoryBot.create(:active_ivl_gold_health_product, hios_id: "11111111122302-01", csr_variant_id: "01")}
       let!(:subject) {Enrollments::IndividualMarket::OpenEnrollmentBegin.new}
 
-
-      # let(:family_health_and_dental)      
+      # let(:family_health_and_dental)
 
       it "the collection should include ten or more Families" do
         # expect(Family.all.size).to be >= 10
@@ -84,43 +84,43 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
         expect(subject.active_enrollment_from_family(enrollment).first).to eq enrollment
       end
 
-      it "at least one Family with both active Individual Market Health and Dental plan Enrollments"
-      it "at least one Family with an active 'Individual Market Health plan Enrollment only'"
-      it "at least one Family with an active 'Assisted Individual Market Health plan Enrollment only'"
-      it "at least one Family with an active 'Individual Market Catastrophic plan Enrollment only'"
-      it "at least one Family with two active Individual Market Health plan Enrollments, one which is responsible person"
-      it "at least one Family with active Individual and SHOP Market Health plan Enrollments"
-      it "at least one Family with active Individual Market Dental and SHOP Market Health plan Enrollments"
-      it "at least one Family with a terminated 'Individual Market Health plan Enrollment only'"
-      it "at least one Family with a terminated 'Individual Market Dental plan Enrollment only'"
-      it "at least one Family with a future terminated 'Individual Market Health plan Enrollment only'"
-      it "at least one Family with a future terminated 'Individual Market Dental-only plan Enrollment'"
+      it "at least one Family with both active Individual Market Health and Dental product Enrollments"
+      it "at least one Family with an active 'Individual Market Health product Enrollment only'"
+      it "at least one Family with an active 'Assisted Individual Market Health product Enrollment only'"
+      it "at least one Family with an active 'Individual Market Catastrophic product Enrollment only'"
+      it "at least one Family with two active Individual Market Health product Enrollments, one which is responsible person"
+      it "at least one Family with active Individual and SHOP Market Health product Enrollments"
+      it "at least one Family with active Individual Market Dental and SHOP Market Health product Enrollments"
+      it "at least one Family with a terminated 'Individual Market Health product Enrollment only'"
+      it "at least one Family with a terminated 'Individual Market Dental product Enrollment only'"
+      it "at least one Family with a future terminated 'Individual Market Health product Enrollment only'"
+      it "at least one Family with a future terminated 'Individual Market Dental-only product Enrollment'"
 
 
       context "and only Families eligible for enrollment auto renewal processing are selected from database" do
 
-        it "the set should include Families with both active Individual Market Health and Dental plan Enrollments"
-        it "the set should include Families with an active 'Individual Market Health plan Enrollment only'"
-        it "the set should include Families with active Individual and SHOP Market Health plan Enrollments"
-        it "the set should include Families with active Individual Market Dental and SHOP Market Health plan Enrollments"
+        it "the set should include Families with both active Individual Market Health and Dental product Enrollments"
+        it "the set should include Families with an active 'Individual Market Health product Enrollment only'"
+        it "the set should include Families with active Individual and SHOP Market Health product Enrollments"
+        it "the set should include Families with active Individual Market Dental and SHOP Market Health product Enrollments"
 
-        it "the set should not include Families with a terminated 'Individual Market Health plan Enrollment only'"
-        it "the set should not include Families with a terminated 'Individual Market Dental plan Enrollment only'"
-        it "the set should not include Families a future terminated 'Individual Market Health plan Enrollment only'"
-        it "the set should not include Families a future terminated Individual Market Dental-only plan Enrollment"
+        it "the set should not include Families with a terminated 'Individual Market Health product Enrollment only'"
+        it "the set should not include Families with a terminated 'Individual Market Dental product Enrollment only'"
+        it "the set should not include Families a future terminated 'Individual Market Health product Enrollment only'"
+        it "the set should not include Families a future terminated Individual Market Dental-only product Enrollment"
 
 
-        context "and the Family with both active Individual Market Health and Dental plan Enrollments is renewed" do
+        context "and the Family with both active Individual Market Health and Dental product Enrollments is renewed" do
           it "should create a new Health enrollment"
-          it "the new enrollment's health plan should be valid for the upcoming calendar year"
+          it "the new enrollment's health product should be valid for the upcoming calendar year"
           it "the new enrollment's effective date should be Jan 1 of next calendar year"
-          it "the new enrollment should include all the enrollees from the current plan year"
+          it "the new enrollment should include all the enrollees from the current product year"
           it "the new enrollment should successfully calculate premium"
 
-          it "should create a new Dental plan enrollment"
-          it "the new enrollment's dental plan should be valid for the upcoming calendar year"
+          it "should create a new Dental product enrollment"
+          it "the new enrollment's dental product should be valid for the upcoming calendar year"
           it "the new enrollment's effective date should be Jan 1 of next calendar year"
-          it "the new enrollment should include all the enrollees from the current plan year"
+          it "the new enrollment should include all the enrollees from the current product year"
           it "the new enrollment should have a calculatable premium"
 
           context "and one child dependent is over age 26 on Jan 1" do
@@ -138,49 +138,49 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
           end
         end
 
-        context "and the Family with an active 'Assisted Individual Market Health plan Enrollment only' is renewed" do
+        context "and the Family with an active 'Assisted Individual Market Health product Enrollment only' is renewed" do
 
           context "and a financial eligibility determination is found" do
             context "and the determination end date is earlier than Jan 1 of the next calendar year" do
-              it "the renewed enrollment should have $0 APTC for the new plan year"
+              it "the renewed enrollment should have $0 APTC for the new product year"
 
-              context "and the renewed enrollment is a silver plan" do
-                it "the renewed plan should not have a CSR variant"
+              context "and the renewed enrollment is a silver product" do
+                it "the renewed product should not have a CSR variant"
               end
             end
 
             context "and the financial assistance redetermination has no end date" do
               it "the renewed enrollment should have the same APTC percentage as the current enrollment"
 
-              context "and the renewed enrollment is a silver plan" do
-                it "the renewed plan should not have the CSR variant from the financial redetermination"
+              context "and the renewed enrollment is a silver product" do
+                it "the renewed product should not have the CSR variant from the financial redetermination"
               end
             end
           end
         end
 
-        context "and the Family with an active 'Individual Market Catastrophic plan Enrollment only' is renewed" do
+        context "and the Family with an active 'Individual Market Catastrophic product Enrollment only' is renewed" do
 
           context "and none of the enrollment members are over age 30 on Jan 1" do
-            it "the new enrollment should have a comparable Catastrophic plan"
+            it "the new enrollment should have a comparable Catastrophic product"
           end
 
           context "and at least one of the enrollment members are over age 30 on Jan 1" do
-            it "should renew enrollment into a mapped Bronze plan??"
+            it "should renew enrollment into a mapped Bronze product??"
           end
         end
 
-        context "and the Family with two active Individual Market Health plan Enrollments, one which is responsible person is renewed" do
+        context "and the Family with two active Individual Market Health product Enrollments, one which is responsible person is renewed" do
           it "should produce a new standard enrollment"
           it "should produce a new responsible person enrollment"
         end
 
-        context "and the Family with active Individual and SHOP Market Health plan Enrollments" do
+        context "and the Family with active Individual and SHOP Market Health product Enrollments" do
           it "should produce a new Individual health enrollment"
           it "should not produce a new SHOP health enrollment"
         end
 
-        context "and the Family with active Individual Market Dental and SHOP Market Health plan Enrollments" do
+        context "and the Family with active Individual Market Dental and SHOP Market Health product Enrollments" do
           it "should produce a new Individual dental enrollment"
           it "should not produce a new SHOP health enrollment"
         end
@@ -197,29 +197,43 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
     end
 
     context "Today is the first day of the Annual Open Enrollment period for the HBX Individual Market" do
-      context "and health and dental plans and rates for the new calendar year are loaded" do
-        context "and comparable plan mapping from this calendar year to the next calendar year is present" do
+      context "and health and dental products and rates for the new calendar year are loaded" do
+        context "and comparable product mapping from this calendar year to the next calendar year is present" do
           context "and Assisted QHP Families have finanancial eligibility redetermined for the next calendar year" do
           end
         end
       end
     end
 
-    describe "To test passive renewals with only ivl health plans" do
+    describe "To test passive renewals with only ivl health products" do
       include_context "setup families enrollments"
-      let!(:subject) {Enrollments::IndividualMarket::OpenEnrollmentBegin.new}
 
       context "Given a database of Families" do
+        it "at least one Family with an active 'Individual Market Health product Enrollment only'" do
+          expect(family_unassisted.active_household.hbx_enrollments.first.kind).to eq "individual"
+        end
+
+        it "at least one Family with an active 'Assisted Individual Market Health product Enrollment only'" do
+          expect(family_assisted.active_household.hbx_enrollments.first.applied_aptc_amount).not_to eq 0
+        end
 
         context "when OE script is executed" do
 
           before do
-            hbx_profile.benefit_sponsorship.benefit_coverage_periods.detect {|bcp| bcp.contains?(renewal_calender_date.beginning_of_year)}.update_attributes!(slcsp_id: renewal_csr_87_plan.id)
+            hbx_profile.benefit_sponsorship.benefit_coverage_periods.each do |bcp|
+              slcsp_id = if bcp.start_on.year == renewal_csr_87_product.application_period.min.year
+                           renewal_csr_87_product.id
+                         else
+                           active_csr_87_product.id
+                         end
+              bcp.update_attributes!(slcsp_id: slcsp_id)
+            end
             hbx_profile.reload
-            renewal_individual_health_plan.reload
-            active_individual_health_plan.reload
+
+            renewal_individual_health_product.reload
+            active_individual_health_product.reload
             family_assisted.active_household.reload
-            allow(Caches::PlanDetails).to receive(:lookup_rate) {|id, start, age| age * 1.0}
+            allow(::BenefitMarkets::Products::ProductRateCache).to receive(:lookup_rate) {|_id, _start, age| age * 1.0}
           end
 
           it "should generate renewal enrollment for assisted family" do
@@ -227,31 +241,34 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
             family_assisted.active_household.reload
             enrollments = family_assisted.active_household.hbx_enrollments
             expect(enrollments.size).to eq 2
-            expect(enrollments[1].applied_aptc_amount.to_f).to eq (enrollments[1].total_premium * enrollments[1].plan.ehb).round(2)
+            expect(enrollments[1].applied_aptc_amount.to_f).to eq((enrollments[1].total_premium * enrollments[1].product.ehb).round(2))
+          end
+
+          it "should generate renewal enrollment for unassisted family" do
+            invoke_oe_script
+            family_unassisted.active_household.reload
+            expect(family_unassisted.active_household.hbx_enrollments.count).to eq 2
           end
         end
       end
     end
 
-    after(:all) do
+    after :all do
       path = "#{Rails.root}/pids/"
-      if Dir.exists?(path)
-        FileUtils.rm_rf(path)
-      end
+      FileUtils.rm_rf(path) if Dir.exist?(path)
     end
   end
 end
 
-
 private
 
 def invoke_oe_script
-  field_names =
-      ["Icnumber", "Ssn", "Subscriber", "Member", "Firstname", "Lastname", "Dob",
-       "#{renewal_calender_year - 1.year} Applied", "#{renewal_calender_year - 1.year} Max", "Unadjustedapplied", "Applied Pct",
-       "#{renewal_calender_year} Aptc", "#{renewal_calender_year} Applied", "#{renewal_calender_year} Csr", "Error Msg"]
-  Dir.mkdir("pids") unless File.exists?("pids")
-  file_name = "#{Rails.root}/pids/2018_FA_Renewals.csv"
+  field_names = ["Icnumber", "Ssn", "Subscriber", "Member", "Firstname", "Lastname", "Dob",
+                 "#{renewal_calender_year - 1.year} Applied", "#{renewal_calender_year - 1.year} Max",
+                 "Unadjustedapplied", "Applied Pct", "#{renewal_calender_year} Aptc",
+                 "#{renewal_calender_year} Applied", "#{renewal_calender_year} Csr", "Error Msg"]
+  Dir.mkdir("pids") unless File.exist?("pids")
+  file_name = "#{Rails.root}/pids/#{renewal_calender_year}_FA_Renewals.csv"
 
   CSV.open(file_name, "w") do |csv|
     csv << field_names
